requirements.txt
flask
flask-cors
pymilvus
sentence-transformers
langchain-google-genai
langchain
pypdf
python-docx
python-dotenv